Refactor VehicleDTOs . Using VehicleStatus for Auto_Vehicles and Auto_VehicleChangeRequest

+++ DTO/VehicleRequest/VehicleEditDTO.cs
@@ -5,7 +5,7 @@
         public string PlateNumber { get; set; } = string.Empty;
         public string Color { get; set; } = string.Empty;
         public string ChassisNumber { get; set; } = string.Empty;
-        public int SeatCount { get; set; }
-        public int DoorCount { get; set; }
+        public byte SeatCount { get; set; }
+        public byte DoorCount { get; set; }
     }
 }
